Webpack!!
Phew, that was a lot of work. Setting up webpack, changing everything
to imports, fixing typescript errors, finding a workaround for Jekyll
templating, reworking the file structure, etc. But in the end, no more
CDNs and the type system is robust!

Also, note that I removed most of the babel stuff, as described in the
README.